Use SwiGLU feed-forward layers
