+++ components/__init__.py
@@ -2,3 +2,4 @@ from .vector_quantizer import VectorQuantizer
 from .learned_query_attention import LearnedQueryAttention
 from .utils import token_entropy, entropy_segments, build_segment_queries_mask
 from .hierarchical_autoencoder import HierarchicalAutoencoder
+from .swiglu import SwiGLU
