+++ components/expander.py
@@ -5,18 +5,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from .rope import RotaryCache, apply_rope
-# SwiGLU Feed Forward
-# ---------------------------------------------------------------------------
-class SwiGLU(nn.Module):
-    def __init__(self, d_model: int, hidden_dim: int):
-        super().__init__()
-        self.w1 = nn.Linear(d_model, hidden_dim * 2)
-        self.w2 = nn.Linear(hidden_dim, d_model)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        a, b = self.w1(x).chunk(2, dim=-1)
-        x = F.silu(a) * b
-        return self.w2(x)
+from .swiglu import SwiGLU
 
 
 # ---------------------------------------------------------------------------
