+++ components/swiglu.py
@@ -1,0 +1,15 @@
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+
+class SwiGLU(nn.Module):
+    """Feed-forward block with SwiGLU activation."""
+    def __init__(self, d_model: int, hidden_dim: int):
+        super().__init__()
+        self.w1 = nn.Linear(d_model, hidden_dim * 2)
+        self.w2 = nn.Linear(hidden_dim, d_model)
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        a, b = self.w1(x).chunk(2, dim=-1)
+        x = F.silu(a) * b
+        return self.w2(x)
